+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added tooltip to "No Active Layers" text
 - Added loading alert on time period change
 - Added last updated date and gage height to AQ popup headers
+- Added opacity sliders for All Features, RT Streamgages, and supporting layers.
 
 ### Changed